Forced calibration to 64 bit variables

+++ INA226.cpp
@@ -34,7 +34,8 @@ void INA226_Class::begin(const uint8_t maxBusAmps, const uint32_t nanoOhmR) { //
       delay(I2C_DELAY);                                                       // Wait for INA to finish resetting //
       if (readWord(INA_CONFIGURATION_REGISTER)==INA_DEFAULT_CONFIGURATION) {  // Yes, we've found an INA226!      //
         _Current_LSB = (uint32_t)maxBusAmps*1000000000/32767;                 // Get the best possible LSB in nA  //
-        _Calibration = 51200000/((int64_t)_Current_LSB*nanoOhmR/100000);      // Compute calibration register     //
+        _Calibration = (int64_t)51200000 / ((int64_t)_Current_LSB *           // Compute calibration register     //
+                       (int64_t)nanoOhmR / (int64_t)100000);                  //                                  //
         _Power_LSB   = (uint32_t)25*_Current_LSB;                             // Fixed multiplier for INA219      //
         writeWord(INA_CALIBRATION_REGISTER,_Calibration);                     // Write the calibration value      //
         break;                                                                // Stop searching                   //
